🐛 fix(Footer.tsx): Uncomment and update allRightsReserved variable 🐛 fix(Hero.tsx): Add null check for image variable 🐛 fix(CapowHead.tsx): Import MtmoLocal component 🐛 fix(all-usps.js): Fix import statement and switch case conditions ✨ feat(Mtmo.js): Add MtmoLocal component 🌐 chore(en.json): Add translation for "All Rights Reserved" 🌐 chore(sv.json): Add translation for "All Rights Reserved"
In Footer.tsx, the commented out code for the allRightsReserved variable is uncommented and updated to use the useIntl hook to format the message based on the locale.

In Hero.tsx, a null check is added for the image variable to prevent errors when it is undefined.

In CapowHead.tsx, the MtmoLocal component is imported to be used in the component.

In all-usps.js, the

🐛 fix(middleware.tsx): remove unused code and fix indentation
✨ feat(middleware.tsx): add language-specific redirects
The unused code for setting the `locale` variable from the `req.cookies` has been removed. The code has been indented properly for better readability. Language-specific redirects have been added to redirect certain paths based on the `locale` variable. For example, if the `locale` is 'en', the paths '/om-oss', '/blogg', and '/kontakt' will be redirected to their English equivalents. Similarly, if the `locale` is 'sv', the paths '/about', '/blog', and '/contact' will be redirected to their Swedish equivalents.

🐛 fix(next.config.js): comment out unused environment variables
The environment variables `MATOMO_DEVELOPMENT_URL` and `MATOMO_PRODUCTION_URL` are currently unused, so they have been commented out.

🐛 fix(index